Changes by PC

+++ src/main/java/com/example/ideasphere/DTOsIN/IndividualCompetitionDTOsIN.java
@@ -40,7 +40,7 @@ public class IndividualCompetitionDTOsIN {
 
     private String status;
 
-    private Integer monetaryReward;
+    private Double monetaryReward;
 
     private Set<Category> categories;
 }
